# docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.CreatePlayerRequest:
    properties:
      height:
        type: number
      jersey_number:
        type: integer
      name:
        type: string
      playstyle:
        type: string
      position:
        type: string
      team_id:
        type: string
      weight:
        type: number
    required:
    - height
    - jersey_number
    - name
    - position
    - team_id
    - weight
    type: object
...
  internal_handler.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
      success:
        type: boolean
    type: object
...
  internal_handler.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  internal_handler.RevertRequest:
    properties:
      target_version:
        type: integer
    required:
    - target_version
    type: object
  internal_handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_handler.UpdatePlayerRequest:
    properties:
      height:
        type: number
      jersey_number:
        type: integer
      name:
        type: string
      playstyle:
        type: string
      position:
        type: string
      version:
        type: integer
      weight:
        type: number
    required:
    - height
    - jersey_number
    - name
    - position
    - version
    - weight
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for managing amateur football teams, players, matches, and
    reports for Perusahaan XYZ.
  title: Kickbase Football Team Management API
  version: "1.0"
paths:
...
  /api/players:
    get:
      description: Get a paginated list of players with optional team, position, name
        filtering and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by team ID
        in: query
        name: team_id
        type: string
      - description: Filter by player name
        in: query
        name: name
        type: string
      - description: Filter by position (CF, SS, CMF, CB, GK, etc.)
        in: query
        name: position
        type: string
      - default: jersey_number
        description: Sort field (name, jersey_number, height, weight, created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.PaginatedResponse'
      summary: List all players
      tags:
      - Players
    post:
      consumes:
      - application/json
      description: Add a new player to a team
      parameters:
      - description: Player data
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreatePlayerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create a new player
      tags:
      - Players
  /api/players/{id}:
    delete:
      description: Soft delete a player (only if no goal records)
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Delete a player
      tags:
      - Players
    get:
      description: Get a single player by its ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get a player by ID
      tags:
      - Players
    put:
      consumes:
      - application/json
      description: Update an existing player's information
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      - description: Player data with version
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdatePlayerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update a player
      tags:
      - Players
  /api/players/{id}/history:
    get:
      description: Get the version history of a player
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
      summary: Get player history
      tags:
      - Players
  /api/players/{id}/revert:
    post:
      consumes:
      - application/json
      description: Revert a player to a specific previous version
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      - description: Target version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RevertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Revert player to previous version
      tags:
      - Players
  /api/reports/matches:
    get:
      description: Get a paginated list of match reports with optional team and result
        status filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by team ID
        in: query
        name: team_id
        type: string
      - description: Filter by match status / outcome (Tim Home Menang, Tim Away Menang,
          Draw, scheduled)
        in: query
        name: status
        type: string
      - default: match_date
        description: Sort field (match_date, created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.PaginatedResponse'
      summary: List all match reports
      tags:
      - Reports
  /api/reports/matches/{id}:
    get:
      description: Get detailed report for a specific match including score, status,
        top scorer, and cumulative wins
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get a single match report
      tags:
      - Reports
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"